# apps/web/src/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "newsdock API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/articles": {
      "get": {
        "summary": "Rest Articles",
        "operationId": "rest_articles_api_articles_get",
        "parameters": [
          {
            "name": "theme",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Theme"
            }
          },
          {
            "name": "domain",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Domain"
            }
          },
          {
            "name": "text",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Text"
            }
          },
          {
            "name": "time_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Time From"
            }
          },
          {
            "name": "time_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Time To"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/articles/{article_id}": {
      "get": {
        "summary": "Rest Article",
        "operationId": "rest_article_api_articles__article_id__get",
        "parameters": [
          {
            "name": "article_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Article Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/health": {
      "get": {
        "summary": "Rest Health",
        "operationId": "rest_health_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthStatus"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalysisRecord": {
        "properties": {
          "agent_name": {
            "type": "string",
            "title": "Agent Name"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": ["agent_name", "payload", "created_at"],
        "title": "AnalysisRecord"
      },
      "ArticleDetail": {
        "properties": {
          "article_id": {
            "type": "string",
            "title": "Article Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain"
          },
          "published_at": {
            "type": "string",
            "format": "date-time",
            "title": "Published At"
          },
          "ingested_at": {
            "type": "string",
            "format": "date-time",
            "title": "Ingested At"
          },
          "slot": {
            "type": "string",
            "title": "Slot"
          },
          "themes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Themes"
          },
          "persons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Persons"
          },
          "orgs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Orgs"
          },
          "tone": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tone"
          },
          "word_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Word Count"
          },
          "analyses": {
            "items": {
              "$ref": "#/components/schemas/AnalysisRecord"
            },
            "type": "array",
            "title": "Analyses"
          }
        },
        "type": "object",
        "required": [
          "article_id",
          "title",
          "url",
          "domain",
          "published_at",
          "ingested_at",
          "slot",
          "themes",
          "persons",
          "orgs",
          "tone",
          "word_count",
          "analyses"
        ],
        "title": "ArticleDetail"
      },
      "ArticleSummary": {
        "properties": {
          "article_id": {
            "type": "string",
            "title": "Article Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain"
          },
          "published_at": {
            "type": "string",
            "format": "date-time",
            "title": "Published At"
          },
          "themes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Themes"
          },
          "scores": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scores"
          }
        },
        "type": "object",
        "required": [
          "article_id",
          "title",
          "url",
          "domain",
          "published_at",
          "themes"
        ],
        "title": "ArticleSummary"
      },
      "FeedPage": {
        "properties": {
          "articles": {
            "items": {
              "$ref": "#/components/schemas/ArticleSummary"
            },
            "type": "array",
            "title": "Articles"
          },
          "next_before": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Before"
          }
        },
        "type": "object",
        "required": ["articles", "next_before"],
        "title": "FeedPage",
        "description": "REST /api/articles response (keyset pagination via next_before)."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthStatus": {
        "properties": {
          "db": {
            "type": "boolean",
            "title": "Db"
          },
          "last_published_slot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Published Slot"
          },
          "kafka": {
            "type": "boolean",
            "title": "Kafka"
          }
        },
        "type": "object",
        "required": ["db", "last_published_slot", "kafka"],
        "title": "HealthStatus"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}